pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bbm25-haystack"
dynamic = ["version"]
description = 'Haystack 2.x In-memory Document Store with Enhanced Efficiency'
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
keywords = ["Document Search", "BM25", "LLM Agent", "RAG", "Haystack"]
authors = [
  { name = "Yuxuan Wang", email = "wangy49@seas.upenn.edu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "haystack-ai",
  "sentencepiece",
  "typing_extensions",
]

[project.urls]
Documentation = "https://github.com/Guest400123064/bbm25-haystack#readme"
Issues = "https://github.com/Guest400123064/bbm25-haystack/issues"
Source = "https://github.com/Guest400123064/bbm25-haystack"

[tool.hatch.version]
path = "src/bbm25_haystack/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.9", "3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/bbm25_haystack tests}"
style = [
  "ruff {args:check .}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff {args:check .} --fix",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
target-version = ["py39"]
line-length = 90
skip-string-normalization = true

[tool.ruff]
target-version = "py39"
line-length = 90
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
  # Ignore usage of `lambda` expression
  "E731",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["bbm25_haystack"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["bbm25_haystack", "tests"]
branch = true
parallel = true
omit = [
  "src/bbm25_haystack/__about__.py",
]

[tool.coverage.paths]
bbm25_haystack = ["src/bbm25_haystack", "*/bbm25-haystack/src/bbm25_haystack"]
tests = ["tests", "*/bbm25-haystack/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
minversion = "6.0"
markers = [
  "unit: unit tests",
  "integration: integration tests"
]

[[tool.mypy.overrides]]
module = [
  "haystack.*",
  "pytest.*"
]
ignore_missing_imports = true
